started checking for balances

+++ modules/api/src/main/java/com/eastbarnetschool/ordermatchingengine/api/repository/UserRepositoryImpl.java
@@ -37,7 +37,7 @@ public class UserRepositoryImpl implements UserRepository {
                 Map.of("username", username),
                 (rs, rowId) -> new User(rs.getObject("user_id", UUID.class), rs.getObject("username", String.class), rs.getObject("created_at", Timestamp.class)));
 
-        Balance balance = balancesService.updateOrCreateBalance(user.getUserId(), "USD", (long) 100000, (long) 0);
+        Balance balance = balancesService.updateOrCreateBalance(user.getUserId(), "USD", 100000, 0);
         ArrayList<Balance> balances = new ArrayList<>();
         balances.add(balance);
